+++ flink-table-store-connector/src/main/java/org/apache/flink/table/store/connector/StatefulPrecommittingSinkWriter.java
@@ -1,0 +1,33 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.table.store.connector;
+
+import org.apache.flink.api.connector.sink2.StatefulSink;
+import org.apache.flink.api.connector.sink2.TwoPhaseCommittingSink;
+import org.apache.flink.table.data.RowData;
+import org.apache.flink.table.store.connector.sink.Committable;
+
+/**
+ * The base interface for file store sink writers.
+ *
+ * @param <WriterStateT> The type of the writer's state.
+ */
+public interface StatefulPrecommittingSinkWriter<WriterStateT>
+        extends StatefulSink.StatefulSinkWriter<RowData, WriterStateT>,
+                TwoPhaseCommittingSink.PrecommittingSinkWriter<RowData, Committable> {}
